centralize remaining ops _foo_method_BAR functions in ops
