+++ pandas/core/sparse/array.py
@@ -14,8 +14,7 @@ from pandas import compat
 from pandas.compat import range
 from pandas.compat.numpy import function as nv
 
-from pandas.core.dtypes.generic import (
-    ABCSparseArray, ABCSparseSeries)
+from pandas.core.dtypes.generic import ABCSparseSeries
 from pandas.core.dtypes.common import (
     _ensure_platform_int,
     is_float, is_integer,
@@ -41,38 +40,6 @@ from pandas.core.indexes.base import _index_shared_docs
 
 
 _sparray_doc_kwargs = dict(klass='SparseArray')
-
-
-def _arith_method_SPARSE_ARRAY(op, name, str_rep=None):
-    """
-    Wrapper function for Series arithmetic operations, to avoid
-    code duplication.
-    """
-
-    def wrapper(self, other):
-        if isinstance(other, np.ndarray):
-            if len(self) != len(other):
-                raise AssertionError("length mismatch: {self} vs. {other}"
-                                     .format(self=len(self), other=len(other)))
-            if not isinstance(other, ABCSparseArray):
-                dtype = getattr(other, 'dtype', None)
-                other = SparseArray(other, fill_value=self.fill_value,
-                                    dtype=dtype)
-            return _sparse_array_op(self, other, op, name)
-        elif is_scalar(other):
-            with np.errstate(all='ignore'):
-                fill = op(_get_fill(self), np.asarray(other))
-                result = op(self.sp_values, other)
-
-            return _wrap_result(name, result, self.sp_index, fill)
-        else:  # pragma: no cover
-            raise TypeError('operation with {other} not supported'
-                            .format(other=type(other)))
-
-    if name.startswith("__"):
-        name = name[2:-2]
-    wrapper.__name__ = name
-    return wrapper
 
 
 def _get_fill(arr):
@@ -864,6 +831,6 @@ def _make_index(length, indices, kind):
 
 
 ops.add_special_arithmetic_methods(SparseArray,
-                                   arith_method=_arith_method_SPARSE_ARRAY,
-                                   comp_method=_arith_method_SPARSE_ARRAY,
-                                   bool_method=_arith_method_SPARSE_ARRAY)
+                                   arith_method=ops._arith_method_SPARSE_ARRAY,
+                                   comp_method=ops._arith_method_SPARSE_ARRAY,
+                                   bool_method=ops._arith_method_SPARSE_ARRAY)
