+++ pandas/core/sparse/series.py
@@ -9,12 +9,10 @@ import numpy as np
 import warnings
 
 from pandas.core.dtypes.missing import isna, notna
-from pandas.core.dtypes.common import is_scalar
 
 from pandas.compat.numpy import function as nv
 from pandas.core.index import Index, _ensure_index, InvalidIndexError
 from pandas.core.series import Series
-from pandas.core.frame import DataFrame
 from pandas.core.internals import SingleBlockManager
 from pandas.core import generic
 import pandas.core.common as com
@@ -23,7 +21,7 @@ import pandas._libs.index as libindex
 from pandas.util._decorators import Appender
 
 from pandas.core.sparse.array import (
-    make_sparse, _sparse_array_op, SparseArray,
+    make_sparse, SparseArray,
     _make_index)
 from pandas._libs.sparse import BlockIndex, IntIndex
 import pandas._libs.sparse as splib
@@ -36,52 +34,6 @@ from pandas.core.sparse.scipy_sparse import (
 _shared_doc_kwargs = dict(axes='index', klass='SparseSeries',
                           axes_single_arg="{0, 'index'}",
                           optional_labels='', optional_axis='')
-
-# -----------------------------------------------------------------------------
-# Wrapper function for Series arithmetic methods
-
-
-def _arith_method_SPARSE_SERIES(op, name, str_rep=None):
-    """
-    Wrapper function for Series arithmetic operations, to avoid
-    code duplication.
-
-    str_rep is not used, but is present for compatibility.
-    """
-
-    def wrapper(self, other):
-        if isinstance(other, Series):
-            if not isinstance(other, SparseSeries):
-                other = other.to_sparse(fill_value=self.fill_value)
-            return _sparse_series_op(self, other, op, name)
-        elif isinstance(other, DataFrame):
-            return NotImplemented
-        elif is_scalar(other):
-            with np.errstate(all='ignore'):
-                new_values = op(self.values, other)
-            return self._constructor(new_values,
-                                     index=self.index,
-                                     name=self.name)
-        else:  # pragma: no cover
-            raise TypeError('operation with {other} not supported'
-                            .format(other=type(other)))
-
-    wrapper.__name__ = name
-    if name.startswith("__"):
-        # strip special method names, e.g. `__add__` needs to be `add` when
-        # passed to _sparse_series_op
-        name = name[2:-2]
-    return wrapper
-
-
-def _sparse_series_op(left, right, op, name):
-    left, right = left.align(right, join='outer', copy=False)
-    new_index = left.index
-    new_name = com._maybe_match_name(left, right)
-
-    result = _sparse_array_op(left.values, right.values, op, name,
-                              series=True)
-    return left._constructor(result, index=new_index, name=new_name)
 
 
 class SparseSeries(Series):
@@ -865,6 +817,6 @@ ops.add_flex_arithmetic_methods(SparseSeries, **ops.series_flex_funcs)
 # overwrite basic arithmetic to use SparseSeries version
 # force methods to overwrite previous definitions.
 ops.add_special_arithmetic_methods(SparseSeries,
-                                   arith_method=_arith_method_SPARSE_SERIES,
-                                   comp_method=_arith_method_SPARSE_SERIES,
+                                   ops._arith_method_SPARSE_SERIES,
+                                   comp_method=ops._arith_method_SPARSE_SERIES,
                                    bool_method=None, force=True)
